+++ src/Screenshot.cpp
@@ -32,6 +32,7 @@
 #include "MainWindow.h"
 #include "SumatraPDF.h"
 #include "Translations.h"
+#include "GlobalHotkeys.h"
 #include "Screenshot.h"
 
 static bool IsAppFrame(HWND hwnd) {
@@ -60,19 +61,6 @@ void InitScreenshotHost() {
     gScreenshotHost.GetOwnerHwnd = GetScreenshotOwnerHwnd;
 }
 
-static bool IsOtherSumatraProcessRunning() {
-    DWORD myPid = GetCurrentProcessId();
-    HWND hwnd = nullptr;
-    while ((hwnd = FindWindowEx(HWND_DESKTOP, hwnd, L"SUMATRA_PDF_FRAME", nullptr)) != nullptr) {
-        DWORD pid;
-        GetWindowThreadProcessId(hwnd, &pid);
-        if (pid != myPid) {
-            return true;
-        }
-    }
-    return false;
-}
-
 // find custom shortcut key string for CmdScreenshot, or empty if none
 static Str FindScreenshotShortcut() {
     // check gSettings->shortcuts first (may have been updated at runtime)
@@ -90,44 +78,6 @@ static Str FindScreenshotShortcut() {
         curr = curr->next;
     }
     return {};
-}
-
-static UINT AccelFVirtToHotkeyMod(BYTE fVirt) {
-    UINT mod = 0;
-    if (fVirt & FALT) {
-        mod |= MOD_ALT;
-    }
-    if (fVirt & FCONTROL) {
-        mod |= MOD_CONTROL;
-    }
-    if (fVirt & FSHIFT) {
-        mod |= MOD_SHIFT;
-    }
-    return mod;
-}
-
-void RegisterScreenshotHotkey(HWND hwnd) {
-    Str shortcut = FindScreenshotShortcut();
-    if (!shortcut) {
-        // don't register global hotkey by default; require explicit
-        // Shortcuts entry (e.g. Key = PrtSc, CmdScreenshot)
-        return;
-    }
-    UINT mod = 0;
-    UINT vk = VK_SNAPSHOT;
-    ACCEL accel{};
-    if (ParseShortcutString(shortcut, accel)) {
-        mod = AccelFVirtToHotkeyMod(accel.fVirt);
-        vk = accel.key;
-    }
-    BOOL ok = RegisterHotKey(hwnd, kScreenshotHotkeyId, mod, vk);
-    if (!ok && !IsOtherSumatraProcessRunning()) {
-        MaybeDelayedWarningNotification(fmt("Couldn't register '%s' global hotkey for taking screenshots", shortcut));
-    }
-}
-
-void UnregisterScreenshotHotkey(HWND hwnd) {
-    UnregisterHotKey(hwnd, kScreenshotHotkeyId);
 }
 
 // --- Set Screenshot Hotkey dialog ---
@@ -314,13 +264,14 @@ void SetHotkeyWnd::DoSet(VirtMouseEvent*) {
     }
     logf("SetHotkeyDoSet: setting screenshot hotkey to '%s'\n", newHotkey);
 
+    TempStr globalKey = str::JoinTemp(StrL("Global "), newHotkey);
     Shortcut* sc = FindScreenshotShortcutEntry();
     if (sc) {
-        str::ReplaceWithCopy(&sc->key, newHotkey);
+        str::ReplaceWithCopy(&sc->key, globalKey);
     } else {
         sc = new Shortcut();
         sc->cmd = str::Dup(StrL("CmdScreenshot"));
-        sc->key = str::Dup(newHotkey);
+        sc->key = str::Dup(globalKey);
         sc->name = {};
         sc->toolbarText = {};
         sc->toolbarSvgIcon = {};
@@ -329,9 +280,7 @@ void SetHotkeyWnd::DoSet(VirtMouseEvent*) {
     }
     SaveSettings();
 
-    for (MainWindow* win : gWindows) {
-        RegisterScreenshotHotkey(win->hwndFrame);
-    }
+    ReRegisterGlobalHotkeys();
     committed = true;
     Close();
 }
@@ -351,6 +300,7 @@ void SetHotkeyWnd::DoRemove(VirtMouseEvent*) {
         curr = curr->next;
     }
     SaveSettings();
+    ReRegisterGlobalHotkeys();
     committed = true;
     Close();
 }
@@ -398,9 +348,7 @@ static void TeardownSetHotkeyWnd() {
     gSetHotkeyWnd = nullptr;
     SetHotkeyWnd::CleanupHook();
     if (!w->committed) {
-        for (MainWindow* win : gWindows) {
-            RegisterScreenshotHotkey(win->hwndFrame);
-        }
+        ReRegisterGlobalHotkeys();
     }
     w->ScheduleDelete();
 }
@@ -421,6 +369,7 @@ bool SetHotkeyWnd::Create(HWND owner) {
     hwndOwner = owner;
     Str existing = FindScreenshotShortcut();
     if (existing) {
+        TrimGlobalPrefix(existing);
         currentHotkey = str::Dup(existing);
     }
 
@@ -505,9 +454,7 @@ void ShowSetScreenshotHotkeyDialog(HWND hwndOwner) {
         TeardownSetHotkeyWnd();
     }
 
-    for (MainWindow* win : gWindows) {
-        UnregisterScreenshotHotkey(win->hwndFrame);
-    }
+    UnregisterGlobalHotkeys(GetGlobalHotkeysHwnd());
 
     auto* wnd = new SetHotkeyWnd();
     wnd->hwndOwner = hwndOwner;
@@ -518,9 +465,7 @@ void ShowSetScreenshotHotkeyDialog(HWND hwndOwner) {
     wnd->SetFont(GetAppFont());
     if (!wnd->Create(hwndOwner)) {
         delete wnd;
-        for (MainWindow* win : gWindows) {
-            RegisterScreenshotHotkey(win->hwndFrame);
-        }
+        ReRegisterGlobalHotkeys();
         return;
     }