+++ src/ShortcutParse.h
@@ -9,6 +9,8 @@
 extern Str (*gShortcutLangCode)();
 
 bool IsValidShortcutString(Str shortcut);
+bool IsGlobalShortcut(Str shortcut);
+bool TrimGlobalPrefix(Str& shortcut);
 
 #if OS_WIN
 bool ParseShortcutString(Str shortcut, ACCEL& accel);